Migrate protein database actions to Protein module
