+++ protein/src/org/labkey/protein/ProteinModule.java
@@ -26,9 +26,7 @@ import org.labkey.api.module.DefaultModule;
 import org.labkey.api.module.ModuleContext;
 import org.labkey.api.pipeline.PipelineService;
 import org.labkey.api.protein.CustomAnnotationSetManager;
-import org.labkey.api.protein.CustomProteinListView;
 import org.labkey.api.protein.ProteinAnnotationPipelineProvider;
-import org.labkey.api.protein.ProteinController;
 import org.labkey.api.protein.ProteinSchema;
 import org.labkey.api.protein.fasta.FastaDbLoader;
 import org.labkey.api.protein.query.CustomAnnotationSchema;
@@ -93,6 +91,8 @@ public class ProteinModule extends DefaultModule
     @Override
     protected void init()
     {
+        // TODO: Merge with protein controller
+        addController("annot", AnnotController.class);
         addController("protein", ProteinController.class);
 
         ProteinUserSchema.register(this);
@@ -106,6 +106,7 @@ public class ProteinModule extends DefaultModule
         PipelineService service = PipelineService.get();
         service.registerPipelineProvider(new ProteinAnnotationPipelineProvider(this));
         UsageMetricsService.get().registerUsageMetrics(getName(), () -> Map.of("hasGeneOntologyData", new TableSelector(ProteinSchema.getTableInfoGoTerm()).exists()));
+        AnnotController.registerAdminConsoleLinks();
     }
 
     @Override
@@ -131,9 +132,19 @@ public class ProteinModule extends DefaultModule
     }
 
     @Override
+    public @NotNull Set<Class> getIntegrationTests()
+    {
+        return Set.of
+        (
+            AnnotController.TestCase.class
+        );
+    }
+
+    @Override
     public @NotNull Set<Class> getUnitTests()
     {
-        return Set.of(
+        return Set.of
+        (
             FastaDbLoader.TestCase.class
         );
     }
